+++ CHANGELOG.md
@@ -2,8 +2,14 @@
 
 ---
 
+### Added
+
+- `auto_fix()` -> Detects and fixes columns with mixed data types And process Numbers strings to float in a DataFrame.
+- `auto_fix` parameter in `process_file()` to detect and data type is wrong and fix it in one time
+
 ### Fixes
 
 - Fixed `File_Format` Parameter set to default default value
+- Fixed Some Bugs
 
 ---